feat: hardware interrupt support and timer
hardware interrupts are now enabled and since the timer is on by
default, a simple handler for printing on the cpu timer has been
implemented.

this has caused a race condition so anything that does write, will
now only write when no interrupt is being triggered. This does
reduce the amount of possible writes per second but removes the
chance of a race condition.

test cases have also been updated with this interrupt lock support.